# packages/vibeship-mind/vibeship_mind-5.0.1.tar.gz/vibeship_mind-5.0.1/docs/specs/openapi.yaml
openapi: 3.1.0
info:
  title: Mind v5 API
  description: |
    Decision intelligence API for AI agents. Mind v5 provides hierarchical temporal memory,
    outcome-weighted retrieval, causal reasoning, and collective intelligence through
    federated pattern learning.

    ## Core Concepts

    - **Memories**: Hierarchical temporal memories with outcome-weighted salience
    - **Decisions**: Tracked decision traces with context snapshots
    - **Outcomes**: Feedback signals that adjust memory salience for future retrieval
    - **Context**: Fused retrieval from vector, graph, and keyword sources

    ## Authentication

    All endpoints require Bearer JWT authentication. Tokens are scoped to a single user
    and include the user_id claim.

    ## Rate Limits

    - Standard: 1000 requests/minute
    - Burst: 10000 requests/minute (short bursts)
    - Memory retrieval: 100 requests/second per user

    ## Versioning

    This API follows semantic versioning. Breaking changes require a major version bump.
    The current version is v1. Use the `Accept-Version` header for version negotiation.
  version: 1.0.0
  contact:
    name: Mind v5 Team
    email: mind-api@example.com
    url: https://docs.mind-v5.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.mind-v5.dev/v1
    description: Production server
  - url: https://staging-api.mind-v5.dev/v1
    description: Staging server
...
security:
  - BearerAuth: []
...
paths:
  /memories/retrieve:
    post:
      operationId: retrieveMemories
      summary: Retrieve context with outcome weighting
      description: |
        Retrieves relevant memories for a given query using multi-source fusion:
        - Vector similarity (Qdrant)
        - Graph traversal (FalkorDB, 2-hop)
        - Keyword/BM25 (PostgreSQL full-text)
        - Temporal recency (decay function)
        - Outcome-weighted salience (learned from feedback)

        Results are fused using Reciprocal Rank Fusion (RRF) and optionally
        reranked with a cross-encoder for maximum relevance.

        **SLO**: 200ms p99 latency
      tags:
        - Memories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetrievalRequest'
            examples:
              basic:
                summary: Basic retrieval
                value:
                  query: "What are the user's communication preferences?"
                  limit: 10
              with_levels:
                summary: Filtered by temporal levels
                value:
                  query: "What projects is the user working on?"
                  limit: 20
                  temporal_levels:
                    - identity
                    - seasonal
                  include_causal: true
              with_session:
                summary: Session-scoped retrieval
                value:
                  query: "What did we discuss earlier?"
                  limit: 5
                  session_id: "550e8400-e29b-41d4-a716-446655440000"
                  temporal_levels:
                    - immediate
                    - situational
      responses:
        '200':
          description: Successfully retrieved memories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrievalResponse'
              examples:
                success:
                  summary: Successful retrieval
                  value:
                    trace_id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                    memories:
                      - memory_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                        content: "User prefers concise, bullet-point responses"
                        temporal_level: identity
                        salience: 0.92
                        outcome_score: 0.15
                        recency_score: 0.3
                        fusion_rank: 1
                      - memory_id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                        content: "Currently working on Mind v5 architecture"
                        temporal_level: seasonal
                        salience: 0.85
                        outcome_score: 0.08
                        recency_score: 0.7
                        fusion_rank: 2
                    causal_edges: []
                    retrieval_latency_ms: 47.3
                    fusion_method: "rrf"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token with the following claims:
        - `sub`: User ID (UUID)
        - `iat`: Issued at timestamp
        - `exp`: Expiration timestamp
        - `scope`: Allowed operations (read, write, admin)
  schemas:
...
    TemporalLevel:
      type: string
      description: |
        Hierarchical temporal level:
        - `immediate`: Current session, working memory (hours)
        - `situational`: Active tasks, recent events (weeks)
        - `seasonal`: Projects, recurring patterns (months)
        - `identity`: Core values, stable preferences (years)
      enum:
        - immediate
        - situational
        - seasonal
        - identity
      example: identity
...
    RetrievalRequest:
      type: object
      description: Request for context retrieval
      required:
        - query
      properties:
        query:
          type: string
          description: Natural language query for retrieval
          maxLength: 2000
          example: "What are the user's communication preferences?"
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Maximum number of memories to return
          example: 10
        temporal_levels:
          type: array
          items:
            $ref: '#/components/schemas/TemporalLevel'
          description: Filter to specific temporal levels
          example:
            - identity
            - seasonal
        session_id:
          type: string
          format: uuid
          description: Scope to specific session
        include_causal:
          type: boolean
          default: false
          description: Include causal edges in response
        min_salience:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          default: 0.0
          description: Minimum effective salience threshold
        rerank:
          type: boolean
          default: true
          description: Apply cross-encoder reranking
    RetrievalResponse:
      type: object
      description: Response containing retrieved memories
      required:
        - trace_id
        - memories
        - retrieval_latency_ms
        - fusion_method
      properties:
        trace_id:
          type: string
          format: uuid
          description: Trace ID for this retrieval (use in decision tracking)
          example: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
        memories:
          type: array
          items:
            $ref: '#/components/schemas/RetrievedMemory'
          description: Retrieved memories ordered by relevance
        causal_edges:
          type: array
          items:
            $ref: '#/components/schemas/CausalEdge'
          description: Related causal edges (if requested)
        retrieval_latency_ms:
          type: number
          format: float
          description: Total retrieval latency in milliseconds
          example: 47.3
        fusion_method:
          type: string
          enum:
            - rrf
            - weighted
            - simple
          description: Fusion method used
          example: rrf
    RetrievedMemory:
      type: object
      description: A memory item with retrieval scores
      required:
        - memory_id
        - content
        - temporal_level
        - salience
        - fusion_rank
      properties:
        memory_id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        content:
          type: string
          example: "User prefers concise, bullet-point responses"
        temporal_level:
          $ref: '#/components/schemas/TemporalLevel'
        salience:
          type: number
          format: float
          description: Effective salience score
          example: 0.92
        outcome_score:
          type: number
          format: float
          description: Contribution from outcome learning
          example: 0.15
        recency_score:
          type: number
          format: float
          description: Temporal recency score
          example: 0.3
        vector_similarity:
          type: number
          format: float
          description: Vector similarity score
          example: 0.87
        fusion_rank:
          type: integer
          description: Final rank after RRF fusion
          example: 1
    CausalEdge:
      type: object
      description: A causal relationship between entities
      required:
        - source_entity
        - target_entity
        - relationship
        - causal_direction
        - causal_strength
      properties:
        source_entity:
          type: string
          description: Source entity name
          example: "deadline_pressure"
        target_entity:
          type: string
          description: Target entity name
          example: "prefer_concise"
        relationship:
          type: string
          description: Relationship type
          example: "causes"
        causal_direction:
          type: string
          enum:
            - causes
            - correlates
            - prevents
          description: Type of causal relationship
          example: causes
        causal_strength:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Strength of causal relationship
          example: 0.73
        conditions:
          type: array
          items:
            type: string
          description: Conditions under which relationship holds
          example:
            - "under_deadline"
            - "low_energy"
        counterfactual:
          type: string
          description: What-if explanation
          example: "Without deadline pressure, user accepts longer explanations"
        confidence:
          type: number
          format: float
          description: Confidence in this edge
          example: 0.87

    # === Decision Schemas ===
...
    Error:
      type: object
      description: Standard error response
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: "MEMORY_NOT_FOUND"
        message:
          type: string
          description: Human-readable error message
          example: "Memory a1b2c3d4-e5f6-7890-abcd-ef1234567890 not found"
        request_id:
          type: string
          description: Request ID for debugging
          example: "req_abc123"
        details:
          type: object
          additionalProperties: true
          description: Additional error context
    ValidationError:
      type: object
      description: Validation error with field details
      required:
        - code
        - message
        - request_id
        - errors
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Request validation failed"
        request_id:
          type: string
          example: "req_abc123"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "query"
              message:
                type: string
                example: "Query must not be empty"
              code:
                type: string
                example: "required"
  responses:
    BadRequest:
      description: Bad request - malformed input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "BAD_REQUEST"
            message: "Request body is not valid JSON"
            request_id: "req_abc123"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "UNAUTHORIZED"
            message: "Invalid or expired authentication token"
            request_id: "req_abc123"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            code: "VALIDATION_ERROR"
            message: "Request validation failed"
            request_id: "req_abc123"
            errors:
              - field: "query"
                message: "Query must not be empty"
                code: "required"
    RateLimited:
      description: Too many requests
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per minute
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "RATE_LIMITED"
            message: "Too many requests. Please retry after 30 seconds."
            request_id: "req_abc123"
            details:
              retry_after: 30
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "INTERNAL_ERROR"
            message: "An unexpected error occurred. Please try again."
            request_id: "req_abc123"